backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.9
click==8.2.0
coloredlogs==15.0.1
cryptography==44.0.3
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
executing==2.2.0
fastapi==0.115.9
fastembed==0.7.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
google-auth==2.40.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.78.1
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
posthog==4.0.1
prompt_toolkit==3.0.51
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.25.5
pypandoc==1.15
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
